refactor(cli): extract design cluster into commands/design.ts (5.0-alpha)
Ninth cut of the cli.ts god-module split (entangled chain: check -> design ->
install -> login -> setup Phase B + review).

- New src/commands/design.ts: cmdDesign (exported) — self-contained a11y +
  design-token gate, baseline-aware.
- cli.ts imports cmdDesign back; the still-inline cmdReview calls the imported
  cmdDesign. COMMANDS table unchanged.
- Dynamic imports re-pathed ./ -> ../ (check-design.js, baseline.js).

cli.ts: 3219 -> 3177 lines (-42; -3153 cumulative from 6330). Typecheck,
eslint (0 errors), full suite (2580), prettier all green; kit design smoke-tested.
